PXForms split to a second project lib (jar)
Cleanup code
