+++ PXForm/src/com/ievolutioned/pxform/PXFEdit.java
@@ -1,4 +1,4 @@
-package com.ievolutioned.iac.pxform;
+package com.ievolutioned.pxform;
 
 import java.util.Map;
 import java.util.Map.Entry;
